ship the image with rubocop
This commit adds [Rubocop][rubocop] to the CircleCI image. This means
that Ruby projects can opt to run the Rubocop linter as part of their
CI build before the need to build their project's Docker image.

To clarify, the linting step can fail when the code is poorly formatted
and with this commit merged, projects can get a faster failure than they
otherwise would have.

If this is accepted, we should ship this image with other common
linters.

Additionally, this commit replaces the ADD instruction in the Dockerfile
with COPY as the former has access to features that we don't need.

[rubocop]: https://github.com/rubocop-hq/rubocop