Allow custom tags for ECR docker image
Currently the ECR docker image is only tagged with the SHA1 of the commit. However, we'd also like to be able to tag it with 'latest' so it's possible for people to pull the latest Identity container for local use. As we're using this helper script to push the container, we need to update it to allow additional tags to be passed in.