ci: explicitly set project name
It seems that docker-compose will set the image name based on the
current directory and the service name instead of "project_app", e.g. if
we use `ci build` in circleci config for "myapp" and build "myservice"
then the image is tagged "myapp_myservice". This means that `ci tag`
cannot find it as it is explicitly looking for "project_app".

This patch just ensures we explicitly set the project name when building
and then look for that image tag when tagging.